Changing how batched primitives are provided for rendering
Using a passed in IEnumerable -- this can be an array, list, or build on the fly using yield returns depending on how it's used.

Just using a static list or array in the primitive class was a little wasteful -- it was doubling the storage requirements as position would also be stored elsewhere.

Was debating having classes using the primitive store pointers to the array inside the class, but that gets messy with dynamic objects. Enemies spawning in and out would trash the array coherency.

Settled on passed in IEnumerable to let whatever is using the rendering decide how to best store/gather the rendering information. Not sure what the overhead is on a foreach w/ yield returns . For truly massive amounts of primitives may want to use static arrays of data that can be directly memcpy'd to GPU?